Split the old GUI class into 2 separate classes + one upper wrapper. It is still not a nice coding style. But at least there are two classes with clearly different functionality and
